Relax requirement for base version
I failed to install erd on Ubuntu 12.04 LTS due to the higher version
of base library required: 4.7 vs 4.5.

After unsuccessful attempts to:
  1. downloading latest version of Haskell Platform
     (no 32 bits version found)
  2. installing latest version of Haskell Platform from source
     (requires higher versions of GHC/cabal than available in Ubuntu 12.04)
  3. installing a higher version of GHC from source
     (requires a higher version of GHC than available in Ubuntu 12.04)

I then thought of changing the requirement in erd itself (which works).

I have set an upper bound for the version to the next major version,
which I have read is a requirement for the package to be accepted
on hackage, and have set the lower bound to the version of the base
package which is installed in Ubuntu 12.04 LTS.

I found this information with:

  cabal info base | less

It reads:

  Versions installed: 4.5.0.0

Note that a lower version of the base library would probably work
as well, and this bound could be lowered further if confirmed.